+++ frontend/components/commons/forms/questions-form/fields/single-label/SingleLabel.component.vue
@@ -1,17 +1,15 @@
 <template>
   <div class="wrapper">
     <QuestionHeaderComponent
-      :title="title"
-      :isRequired="isRequired"
-      :hasSuggestion="hasSuggestion"
-      :tooltipMessage="description"
+      :question="question"
+      :showSuggestion="showSuggestion"
     />
 
     <LabelSelectionComponent
-      v-model="uniqueOptions"
-      :multiple="false"
-      :componentId="questionId"
+      v-model="question.answer.values"
+      :componentId="question.id"
       :maxOptionsToShowBeforeCollapse="maxOptionsToShowBeforeCollapse"
+      :multiple="false"
     />
   </div>
 </template>
@@ -20,54 +18,18 @@
 export default {
   name: "SingleLabelComponent",
   props: {
-    questionId: {
-      type: String,
+    question: {
+      type: Object,
       required: true,
     },
-    title: {
-      type: String,
-      required: true,
-    },
-    options: {
-      type: Array,
-      required: true,
-    },
-    isRequired: {
+    showSuggestion: {
       type: Boolean,
       default: () => false,
     },
-    description: {
-      type: String,
-      default: () => "",
-    },
-    visibleOptions: {
-      type: Number | null,
-      required: false,
-    },
-    hasSuggestion: {
-      type: Boolean,
-      default: () => false,
-    },
-  },
-  model: {
-    prop: "options",
-  },
-  data() {
-    return {
-      uniqueOptions: [],
-    };
-  },
-  beforeMount() {
-    this.uniqueOptions = this.options.reduce((accumulator, current) => {
-      if (!accumulator.find((item) => item.id === current.id)) {
-        accumulator.push(current);
-      }
-      return accumulator;
-    }, []);
   },
   computed: {
     maxOptionsToShowBeforeCollapse() {
-      return this.visibleOptions ?? -1;
+      return this.question.settings.visible_options ?? -1;
     },
   },
 };
@@ -77,6 +39,6 @@ export default {
 .wrapper {
   display: flex;
   flex-direction: column;
-  gap: 12px;
+  gap: $base-space;
 }
 </style>