+++ frontend/playwright.config.ts
@@ -31,12 +31,6 @@ export default defineConfig({
     trace: "on-first-retry",
   },
 
-  expect: {
-    toHaveScreenshot: {
-      maxDiffPixelRatio: 0.2,
-    },
-  },
-
   /* Configure projects for major browsers */
   projects: [
     {